--- a/dependency_range_tests/scripts/with_standard_tests/anthropic/node/update_resolutions_lowest.js
+++ b/dependency_range_tests/scripts/with_standard_tests/anthropic/node/update_resolutions_lowest.js
@@ -1,11 +1,17 @@
 const fs = require("fs");
 const semver = require("semver");
 
-const communityPackageJsonPath = "/app/monorepo/libs/langchain-anthropic/package.json";
+const communityPackageJsonPath =
+  "/app/monorepo/libs/langchain-anthropic/package.json";
 
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
-if (currentPackageJson.peerDependencies["@langchain/core"] && !currentPackageJson.peerDependencies["@langchain/core"].includes("rc")) {
+if (
+  currentPackageJson.peerDependencies?.["@langchain/core"] &&
+  !currentPackageJson.peerDependencies["@langchain/core"].includes("rc")
+) {
   const minVersion = semver.minVersion(
     currentPackageJson.peerDependencies["@langchain/core"]
   ).version;
@@ -15,8 +21,11 @@ if (currentPackageJson.peerDependencies["@langchain/core"] && !currentPackageJso
   };
 }
 
-if (currentPackageJson.devDependencies["@langchain/core"]) {
+if (currentPackageJson.devDependencies?.["@langchain/core"]) {
   delete currentPackageJson.devDependencies["@langchain/core"];
 }
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);