--- a/dependency_range_tests/scripts/with_standard_tests/anthropic/node/update_workspace_deps.js
+++ b/dependency_range_tests/scripts/with_standard_tests/anthropic/node/update_workspace_deps.js
@@ -1,7 +1,10 @@
 const fs = require("fs");
 
-const communityPackageJsonPath = "/app/monorepo/libs/langchain-anthropic/package.json";
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const communityPackageJsonPath =
+  "/app/monorepo/libs/langchain-anthropic/package.json";
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
 // Anthropic has other workspaces as devDependencies, but tagged as `workspace:*` for the version.
 // Update these to be `latest` for the test.
@@ -12,4 +15,7 @@ if (currentPackageJson.devDependencies["@langchain/community"]) {
   };
 }
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);