--- a/dependency_range_tests/scripts/with_standard_tests/community/node/update_resolutions_latest.js
+++ b/dependency_range_tests/scripts/with_standard_tests/community/node/update_resolutions_latest.js
@@ -1,7 +1,10 @@
 const fs = require("fs");
 
-const communityPackageJsonPath = "/app/monorepo/libs/langchain-community/package.json";
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const communityPackageJsonPath =
+  "/app/monorepo/libs/langchain-community/package.json";
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
 if (currentPackageJson.devDependencies["@langchain/core"]) {
   delete currentPackageJson.devDependencies["@langchain/core"];
@@ -13,4 +16,7 @@ if (currentPackageJson.dependencies["@langchain/openai"]) {
   currentPackageJson.dependencies["@langchain/openai"] = "latest";
 }
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);