--- a/dependency_range_tests/scripts/with_standard_tests/community/node/update_resolutions_lowest.js
+++ b/dependency_range_tests/scripts/with_standard_tests/community/node/update_resolutions_lowest.js
@@ -1,11 +1,17 @@
 const fs = require("fs");
 const semver = require("semver");
 
-const communityPackageJsonPath = "/app/monorepo/libs/langchain-community/package.json";
+const communityPackageJsonPath =
+  "/app/monorepo/libs/langchain-community/package.json";
 
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
-if (currentPackageJson.peerDependencies["@langchain/core"] && !currentPackageJson.peerDependencies["@langchain/core"].includes("rc")) {
+if (
+  currentPackageJson.peerDependencies?.["@langchain/core"] &&
+  !currentPackageJson.peerDependencies["@langchain/core"].includes("rc")
+) {
   const minVersion = semver.minVersion(
     currentPackageJson.peerDependencies["@langchain/core"]
   ).version;
@@ -15,11 +21,14 @@ if (currentPackageJson.peerDependencies["@langchain/core"] && !currentPackageJso
   };
 }
 
-if (currentPackageJson.devDependencies["@langchain/core"]) {
+if (currentPackageJson.devDependencies?.["@langchain/core"]) {
   delete currentPackageJson.devDependencies["@langchain/core"];
 }
 
-if (currentPackageJson.dependencies["@langchain/openai"] && !currentPackageJson.dependencies["@langchain/openai"].includes("rc")) {
+if (
+  currentPackageJson.dependencies?.["@langchain/openai"] &&
+  !currentPackageJson.dependencies["@langchain/openai"].includes("rc")
+) {
   const minVersion = semver.minVersion(
     currentPackageJson.dependencies["@langchain/openai"]
   ).version;
@@ -29,8 +38,11 @@ if (currentPackageJson.dependencies["@langchain/openai"] && !currentPackageJson.
   };
 }
 
-if (currentPackageJson.devDependencies["@langchain/openai"]) {
+if (currentPackageJson.devDependencies?.["@langchain/openai"]) {
   delete currentPackageJson.devDependencies["@langchain/openai"];
 }
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);