--- a/dependency_range_tests/scripts/with_standard_tests/community/node/update_resolutions_npm.js
+++ b/dependency_range_tests/scripts/with_standard_tests/community/node/update_resolutions_npm.js
@@ -1,10 +1,16 @@
 const fs = require("fs");
 
-const communityPackageJsonPath = "/app/monorepo/libs/langchain-community/package.json";
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const communityPackageJsonPath =
+  "/app/monorepo/libs/langchain-community/package.json";
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
 if (currentPackageJson.devDependencies) {
   delete currentPackageJson.devDependencies;
 }
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);