--- a/dependency_range_tests/scripts/with_standard_tests/google-vertexai/node/update_resolutions_latest.js
+++ b/dependency_range_tests/scripts/with_standard_tests/google-vertexai/node/update_resolutions_latest.js
@@ -1,7 +1,10 @@
 const fs = require("fs");
 
-const communityPackageJsonPath = "/app/monorepo/libs/langchain-google-vertexai/package.json";
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const communityPackageJsonPath =
+  "/app/monorepo/libs/langchain-google-vertexai/package.json";
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
 if (currentPackageJson.devDependencies["@langchain/core"]) {
   delete currentPackageJson.devDependencies["@langchain/core"];
@@ -12,4 +15,7 @@ if (currentPackageJson.dependencies["@langchain/google-gauth"]) {
   delete currentPackageJson.dependencies["@langchain/google-gauth"];
 }
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);