--- a/dependency_range_tests/scripts/with_standard_tests/google-vertexai/node/update_resolutions_lowest.js
+++ b/dependency_range_tests/scripts/with_standard_tests/google-vertexai/node/update_resolutions_lowest.js
@@ -1,11 +1,17 @@
 const fs = require("fs");
 const semver = require("semver");
 
-const communityPackageJsonPath = "/app/monorepo/libs/langchain-google-vertexai/package.json";
+const communityPackageJsonPath =
+  "/app/monorepo/libs/langchain-google-vertexai/package.json";
 
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
-if (currentPackageJson.peerDependencies["@langchain/core"] && !currentPackageJson.peerDependencies["@langchain/core"].includes("rc")) {
+if (
+  currentPackageJson.peerDependencies?.["@langchain/core"] &&
+  !currentPackageJson.peerDependencies["@langchain/core"].includes("rc")
+) {
   const minVersion = semver.minVersion(
     currentPackageJson.peerDependencies["@langchain/core"]
   ).version;
@@ -15,11 +21,14 @@ if (currentPackageJson.peerDependencies["@langchain/core"] && !currentPackageJso
   };
 }
 
-if (currentPackageJson.devDependencies["@langchain/core"]) {
+if (currentPackageJson.devDependencies?.["@langchain/core"]) {
   delete currentPackageJson.devDependencies["@langchain/core"];
 }
 
-if (currentPackageJson.dependencies["@langchain/google-gauth"] && !currentPackageJson.dependencies["@langchain/google-gauth"].includes("rc")) {
+if (
+  currentPackageJson.dependencies?.["@langchain/google-gauth"] &&
+  !currentPackageJson.dependencies["@langchain/google-gauth"].includes("rc")
+) {
   const minVersion = semver.minVersion(
     currentPackageJson.dependencies["@langchain/google-gauth"]
   ).version;
@@ -29,4 +38,7 @@ if (currentPackageJson.dependencies["@langchain/google-gauth"] && !currentPackag
   };
 }
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);