--- a/dependency_range_tests/scripts/with_standard_tests/node/update_workspace_dependencies.js
+++ b/dependency_range_tests/scripts/with_standard_tests/node/update_workspace_dependencies.js
@@ -1,8 +1,11 @@
 const fs = require("fs");
 
-const standardTestsPackageJsonPath = "/app/monorepo/libs/langchain-standard-tests/package.json";
+const standardTestsPackageJsonPath =
+  "/app/monorepo/libs/langchain-standard-tests/package.json";
 
-const currentPackageJson = JSON.parse(fs.readFileSync(standardTestsPackageJsonPath));
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(standardTestsPackageJsonPath)
+);
 
 if (currentPackageJson.dependencies["@langchain/core"]) {
   currentPackageJson.dependencies = {
@@ -18,4 +21,7 @@ if (currentPackageJson.devDependencies["@langchain/scripts"]) {
   };
 }
 
-fs.writeFileSync(standardTestsPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  standardTestsPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);