--- a/dependency_range_tests/scripts/with_standard_tests/openai/node/update_resolutions_latest.js
+++ b/dependency_range_tests/scripts/with_standard_tests/openai/node/update_resolutions_latest.js
@@ -1,11 +1,17 @@
 const fs = require("fs");
 
-const communityPackageJsonPath = "/app/monorepo/libs/langchain-openai/package.json";
-const currentPackageJson = JSON.parse(fs.readFileSync(communityPackageJsonPath));
+const communityPackageJsonPath =
+  "/app/monorepo/libs/langchain-openai/package.json";
+const currentPackageJson = JSON.parse(
+  fs.readFileSync(communityPackageJsonPath)
+);
 
 if (currentPackageJson.devDependencies["@langchain/core"]) {
   delete currentPackageJson.devDependencies["@langchain/core"];
   currentPackageJson.peerDependencies["@langchain/core"] = "latest";
 }
 
-fs.writeFileSync(communityPackageJsonPath, JSON.stringify(currentPackageJson, null, 2));
+fs.writeFileSync(
+  communityPackageJsonPath,
+  JSON.stringify(currentPackageJson, null, 2)
+);