--- a/docs/core_docs/static/js/job_link.js
+++ b/docs/core_docs/static/js/job_link.js
@@ -1,4 +1,4 @@
-document.addEventListener("DOMContentLoaded", function() {
+document.addEventListener("DOMContentLoaded", function () {
   const jobLink = document.getElementById("js_job_link");
   if (jobLink) {
     const jobUrl = new URL(jobLink.href);