--- a/environment_tests/test-exports-bun/scripts/combine-dependencies.js
+++ b/environment_tests/test-exports-bun/scripts/combine-dependencies.js
@@ -1,8 +1,13 @@
 import * as fs from "fs";
 
-const langchainPackageJson = JSON.parse(fs.readFileSync("../langchain/package.json"));
+const langchainPackageJson = JSON.parse(
+  fs.readFileSync("../langchain/package.json")
+);
 const testPackageJson = JSON.parse(fs.readFileSync("./package.json"));
 
-testPackageJson.dependencies = { ...testPackageJson.dependencies, ...langchainPackageJson.dependencies };
+testPackageJson.dependencies = {
+  ...testPackageJson.dependencies,
+  ...langchainPackageJson.dependencies,
+};
 
 fs.writeFileSync("./package.json", JSON.stringify(testPackageJson, null, 2));