--- a/environment_tests/test-exports-vercel/src/styles/Home.module.css
+++ b/environment_tests/test-exports-vercel/src/styles/Home.module.css
@@ -97,7 +97,7 @@
 
 .center::before,
 .center::after {
-  content: '';
+  content: "";
   left: 50%;
   position: absolute;
   filter: blur(45px);
@@ -125,7 +125,7 @@
 
 .thirteen::before,
 .thirteen::after {
-  content: '';
+  content: "";
   position: absolute;
   z-index: -1;
 }